Pedir turnos (#17)
* Agregando emailSender para el envio de emails

* Agregando un threadpool aparte para ejecutar el envio de emails en distintos threads

* Agrego clase Turno con sus componentes y Cliente. Agrego tests de comportamiento sobre Turno.

* Finalizo la parte del back con los endpoints para pedir y confirmar turno (con el envio de email en este ultimo)

* Modifico query turnoRepository que tenia bug. Ahora el confirmar turno setea al peluquero en OCUPADO. Agrego endpoint finalizar turno y cuando el peluquero no tenga turnos confirmados, este pasa a ser DISPONIBLE de nuevo.

* Agrego endpoint para getear a peluquero por id

* Agrego requestbody annotation en los Post y valid en get peluquero por id

* El precio del turno tiene como base siempre el corteMin. Fix tests turno precio

* Update README.md como instalar y ejecutar el backend

+++ src/main/kotlin/ar/edu/unq/peluqueriayabackend/persistence/GenericDAO.kt
@@ -3,9 +3,9 @@ package ar.edu.unq.peluqueriayabackend.persistence
 import java.util.Optional
 
 interface GenericDAO<T> {
-    fun get(id:Int): Optional<T>
+    fun get(id:Long): Optional<T>
     fun getAll():Collection<T>
     fun save(t:T):T
     fun update(t:T):T
-    fun delete(id:Int)
+    fun delete(id:Long)
 }
